requirement.txt
langchain
groq
python-dotenv
pypdf
faiss-cpu
langchain-groq